+++ transcriptic/config.py
@@ -717,6 +717,7 @@ class Connection(object):
         title=None,
         test_mode=False,
         payment_method_id=None,
+        predecessor_id=None,
     ):
         """Submit specified launch request"""
         payload = {
@@ -725,6 +726,7 @@ class Connection(object):
             "protocol_id": protocol_id,
             "test_mode": test_mode,
             "payment_method_id": payment_method_id,
+            "predecessor_id": predecessor_id,
         }
         data = {k: v for k, v in payload.items() if v is not None}
         return self.post(